+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 2018-09-18 v2.3.0
 
 * Add "Insert or Update (Upsert) a Record Using an External ID" operation, using the SObject Rows by External ID resource. For details on usage see https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/resources_sobject_upsert.htm
-* Minor breaking change: Changed IssuedAt property of AccessTokenResponse from string to int? since the value is a Unix epoch datetime. Added IssuedAtDateTime property for easy date conversion.
+* Minor breaking change: Changed IssuedAt property of AccessTokenResponse from string to long? since the value is a Unix epoch datetime. Added IssuedAtDateTime property for easy date conversion.
 
 ### 2018-04-17 v2.2.0